settings moves to container
